debug: generic latency scripts.
While it makes sense to provide subsystem-specific scripts that will
provide detailed information about what is going on with that code,
there are also times in which we need some ad-hoc way to peek into
latencies between 2 points.

We can easily do that by writing a quick script from a template. But so
far, our probe points are usually called <name>_start and <name>_end,
and we can leverage that to come up with a generic mechanism that given
two probe points that look like that will produce latency figures
between them.

That is extremely useful as a first look at the system.

Signed-off-by: Glauber Costa <glauber@scylladb.com>