docker: only run setup scripts during first boot
We can detect whether or not this is the first boot by checking for the
existence of the docker conf file inside /etc.

The setup scripts shouldn't be run during docker start. Most of them are
harmless if ran more than once, but scylla_io_setup can take many
minutes

Fixes #2112

Signed-off-by: Glauber Costa <glauber@scylladb.com>