docker: move yaml to a different location
Users sometimes need to run their own yaml configuration files, and it
is currently annoying to deploy modified files on docker.

One possible solution is to bind mount the file into the docker
container using the -v switch, just like we already do for for the data
volume.

The problem with the aforementioned approach is that we have to change
the yaml file to insert the addresses, and that will change the file in
the host (or fail to happen, if we bind mount it read-only).

The solution I am proposing is to copy /etc/scylla/scylla.yaml to an
alternative location, and point scylla to that location instead. This
way, a user can still bind mount it (preferrably read only) their
standard yaml file to /etc/scylla/scylla.yaml, but we will operate on
a copy.

If no file is specified, we will just copy the default.

Fixes #2113

Signed-off-by: Glauber Costa <glauber@scylladb.com>