streaming: move mutation limiter to memory-based instead of magic number
We currently restrict the number of in-flight mutations being sent to
the remote node to 256. By doing that, we avoid keeping too much data in
memory.

However, whereas 256 x 1MB mutations hold 256MB of memory, 256 x 100
bytes mutations barely holds 25kB and if we can read more, we could send
more. A better way to handle this is to make this semaphore size-based,
and allow for a maximum amount of memory.

We will allow for 512MB of memory to be used across all shards, with a
maximum of 5 % of shard's memory. This is set so to be enough to
saturate a 40Gbps connection at 100ms latencies

Signed-off-by: Glauber Costa <glauber@scylladb.com>