database: implement API to enable/disable auto compactions
In Scylla, we strive to provide good isolation between foreground and
background jobs, and to adjust shares in the various schedulers present
in the system so things balance themselves out without human
intervention.

However, there are still cases in which a user may be justified in
wanting to disable auto compactions: be it to circumnvent bad behavior
that we may have - as much as we try not to, be it because they have a
seasonal workload that have to handle extreme peaks of writes followed
by quiet times later.

This patch adds the ability to do so.

Signed-off-by: Glauber Costa <glauber@scylladb.com>