compaction_strategy: implement backlog manager for the SizeTiered strategy
The SizeTiered backlog for a single SSTable is defined as:

   Bi = Ei * log4(T / Si)

Where:

  - Si is the size of this individual SSTable
  - T is the sum of sizes for all individual SSTables
  - Ei is the effective bytes in this SSTable.

The Effective size of an SSTable is:
 - The uncompacted size for an SSTable under compaction
 - The partially written size for an SSTable being written
 - The SSTable size for an SSTable that is not undergoing
   any of those processes.

The Aggregate Backlog for the entire Table is just the sum of
all individual SSTable backlogs, including the SSTables currently
being written.

Care is taken to avoid iterating over all SSTables, by separating
the aggregate backlog into a static component (sstables not changing) and
a component of SSTables that are undergoing change.

Signed-off-by: Glauber Costa <glauber@scylladb.com>