compaction_strategy: implement backlog manager for the SizeTiered strategy
The SizeTiered backlog for a single table is defined as:

   Bi = Ei * log4(T / Si)

Where:

  - Si is the size of this individual SSTable
  - T is the sum of sizes for all individual SSTables
  - Ei is the effective bytes in this SSTable, or in
    other words, the SSTable total size minus the bytes
    it already compacted.

The Aggregate Backlog for the entire Table is just the sum of
all individual backlogs.

Care is taken to avoid iterating over all SSTables, by separating
the aggregate backlog into a static component (sstables not changing) and
a component of SSTables that are undergoing change.

Signed-off-by: Glauber Costa <glauber@scylladb.com>