row_cache: modernize use of seastar threads
For a while now we have an async() function, that simplifies the code by not
needing to issue an explicit join. This patch converts the row cache to use
async() as well, which most of our code already does. Doing so will make
it easier to make changes to update_cache.

While working on it, it becomes clear that we have two cleanup functions:
one outside the seastar thread and one inside it. They are basically doing
the same thing and can be merged if we move the code that transfer ownership
inside the thread.

Signed-off-by: Glauber Costa <glauber@scylladb.com>