mutation_reader: keep a copy, not a reference of the restricted reader config
We want to move to per-request timeouts, which means each request will
potentially specify a different timeout setting. That means that we
can't just hold a global reference.

While we could still hold a reference in the mutation_reader and keep a
copy in the outer layer, there isn't really a better way to put it, as
the mutation reader is naturally aligned with the request lifetime.

Signed-off-by: Glauber Costa <glauber@scylladb.com>