pass read timeouts down to the mutation reader layer
Right now we have no way to communicate expectation of timeouts to the
reader layer. The restricted reader itself can timeout based on the
config structure it is passed, but changing this value presents a
challenge.

Starting from the database layer, we need to pass the timeout somehow
to make_reader() and from there to make_sstable_reader. But those
functions are called from the mutation source, which means the mutation
source has to be somehow aware of what the desired timeout is.

One mutation source can serve many reads. So we want that timeout to
be passed not to the mutation source itself, but to the operator(),
which is what creates a reader that will serve a single read.

The bulk of this patch is just patching definitions and callers to
adjust to that new API. Usually for things like this we just add an
optional  parameter to the end of the existing parameter list to avoid
clutter but in this case I am not doing that. The reason is, there are
so many callers that don't specify the other parameters - which now
would have to - that this would generate clutter all the same.

After this patch, the timeout is specified in the query() entry-point.
It is just the default timeout that we would get from the restricted
reader. In a follow up patch we will change that.

Also, reads from cache and memtables will never timeout - same as
before. But going forward, there is now the infrastructure in place
to make that happen if we so want.

Signed-off-by: Glauber Costa <glauber@scylladb.com>