database: pass a restricted reader config as a parameter to make_sstable_reader
Currently we try to determine which restricted reader config to use
based on the I/O priority that is passed down. This is a less than
elegant solution in itself, but as we want to move to per-request
timeouts it gets even worse.

We can either add a timeout parameter and then overwrite the original
timeout, or we can pass the config struct from the upper layer already
the way we want to use: we then void the need of guessing which config
structure to use, because reads and streaming will come from different
places.

Signed-off-by: Glauber Costa <glauber@scylladb.com>