database: push read_concurrency_config yet another layer up
In our quest to achieve per-request timeouts, we have to somehow
communicate the timeout information to column_family::make_reader.
That, in turn requires column_family::as_mutation_source, which is the
caller of make_reader to be aware of the timeout as well.

There are two choices to be made here:

1) will we change the mutation_source constructor or pass the config
through the lambda capture? I have decided for the latter, since I don't
think the mutation_source API should be concerned about that.

2) Will we pass timeouts or the read_concurrency_config at this point of
   the API?
I have also decided for the latter, since it is not unreasonable to
think that we'll also want to play with other parameters (such as the
concurrency) at the make_reader level.

In upcoming patches we'll glue the storage proxy calls into this: at
that point we'll be passing down timeouts only.

Signed-off-by: Glauber Costa <glauber@scylladb.com>