restricted reader: use time_point based timeouts instead of delta
Right now using delta-based timeouts in the restricted reader poses a
problem, in that it is hard to communicate timeouts from upper layers to
the readers. The reason being that timeouts in the upper layer are
usually expressed as lowres_clocks, which we can't easily add to a
nanosecond duration.

Of course, the config structure is not per-request so we can't get
the timeout directly from that structure. It needs to be passed down
the API, and the caller will then get the timeout from the structure
and add it to now() -- a convenience method is added for that.

Moreover, in the interest of having per-request timeouts we also want
to be able to allow the reader anyway. In future patches we will allow
the caller to pass timeouts other than what we have in the configuration
struct.

Signed-off-by: Glauber Costa <glauber@scylladb.com>