sstables: pass references instead of pointers for write_monitor
This came from Avi's review on the read_monitors. He suggests we
wouldn't keep shared pointers, and would instead have the caller
ensuring lifetime. That makes sense, but having the writer interface
using shared_ptr and the read interface using references would lead to
an inconsistent interface.

For the sake of consistency we will change the write monitor to take
references before we do that. From database.cc's perspective, we could
now keep the monitors in a do_with() block, but we will keep the
shared_ptrs to manage their lifetime in anticipation of upcoming patches
in this series, where we'll have to pass them somewhere else.

We are still using pointers in the sstable_writer, because the config
struct is marked as const. As the monitor methods may manipulate data,
its methods can't be const and this will force a copy - not a good idea
to its polymorphic nature.

Signed-off-by: Glauber Costa <glauber@scylladb.com>