mutation_reader_test: fix failing test.
Raphael recently caught this test failing. I can't really reproduce it,
but it seems to me that it is a timing issue: we execute two different
statements, each one should timeout after 10ms. After 20ms, we make sure
that they both timed out.

They don't (in his system), which is very likely explained by the fact
that we are no longer using high resolution clocks for the timeouts,
plus the fact that we are not passing the timeout value down a more
complex path.

So the fix I am proposing here is to just account for potential
innacuracies in the clocks and calculations by waiting a bit longer.

Ideally, we would use the manual clock for this. But in this case, this
would mean adding template parameters to pretty much all of the
mutation_reader path.

Currently, not only the test failed, it also had an use-after-free
SIGSEGV. That happens because we give up on the reader while the
timeouts is still to happen.

It is the caller responsibility to ensure the lifetime of the reader is
correct. So we'll just wait for the future regardless of the result of
the test.

Signed-off-by: Glauber Costa <glauber@scylladb.com>