controllers: fix upper limit for memtable I/O controller
The controller exports the current threshold as a proportion of the
total dirty memory. The CPU controller uses twice the size of the
throttle threshold, which is wrong - it effectively turns the
controller into a real dirty instead of virtual dirty controller.

This is a regression introduced when I reworked the controller code
to introduce the I/O controller. that factor of 2 used to be compensated
in the values and now it isn't.

The I/O controller does it right.
Fix it so that both controllers look like each other.

Signed-off-by: Glauber Costa <glauber@scylladb.com>