controllers: fix upper limit for memtable I/O controller
The controller exports the current threshold as a proportion of the
total dirty memory. That threshold is twice the size of the virtual
limit.

The CPU controller does it right, but the I/O controller doesn't. As a
result, shares will end up higher than they should.

Fix it so that both controllers look like each other.

Signed-off-by: Glauber Costa <glauber@scylladb.com>