abstract summary entry version of the token with a token view
dht::token doesn't have a trivial destructor, so destroying an array
full of those can be quite expensive. If we use the same trick as we
used for the summary; storing the token data in a stable memory
location, we can leave the entries with a trivial destructor and destroy
the chunks themselves. Those being larger, they will be more efficient
to delete.

To avoid messing up with dht::token and adding a view there, I am
keeping token_view as an SSTable class. Before we use it, we convert it
to an actual token: this way we don't need to change the dht methods
everywhere to take a view as well. For the murmur3 partitioner, creating
a dht::token from its token_view should be easy enough - as the token is
8-byte long. The other partitioners don't matter much.

Note that as a requirement, we now need to keep the token alive when
using it: so the decorated_key_view becomes more than just a view - it
actually holds a token. Hopefully that's ok for now.

Signed-off-by: Glauber Costa <glauber@scylladb.com>