add a token_view
Ideally we would like tokens to be trivially destructible, so that we
can easily dispose of giant vectors holding them. While that is hard to
do with our current infrastructure, we can introduce a token_view, which
holds a bytes_view elements instead of the real data - making it
trivially destructible.

The comparators are then changed to take a token_view, and an implicit
conversion function is provided from tokens so they get compared.

token_view is made to be a nested class inside token because otherwise
it gets hard to deal with token::kind - using that in token_view would
create circular dependencies.

Signed-off-by: Glauber Costa <glauber@scylladb.com>