database: move writes to their own scheduling group
When we introduced the CPU scheduler, we have also introduced a group
for commitlog - but never used it. This patch renames it to "writes" and
wraps the entire replica write path in it (both memtable apply and
commitlog manipulation).

The reason for that choice of naming is that unlike I/O - which only
happens for commitlog in the latency path (the actual flush is in the
background), for CPU we have both commitlog and memtable manipulation in
the latency path.

Signed-off-by: Glauber Costa <glauber@scylladb.com>