don't hold SSTable lock while flushing or compacting SSTable
When we flush or compact SSTables, we hold the SSTable lock for reads.
The write part of this lock is taken by the storage service, when serving
the nodetool refresh API.

The reason we do that is that we want to maintain the directory safety
against concurrent changes from new SSTables. The problem with that is
that compactions can run for hours, and during that time nodetool
refresh becomes not really available. As that is a common recovery
operation, it is important that it is fast and reliable.

Many things were suggested to work around this, but one thing we should
note is that there is no need to serialize refresh with data actually
being written to the file: we only need to serialize against file
creation.

So we only really need to hold the lock for time enough for us to create
a generation. Once the column family is aware of that generation we can
just not touch the files contained in that generation, as we would do
for any other existing SSTable. For nodetool refresh, all generations
after the ones already committed for creation are kept available.

Fixes #3064
Tests: unit (release)
       manually verified that nodetool refresh can proceed in parallel
       with an existing long-lived nodetool compact

Signed-off-by: Glauber Costa <glauber@scylladb.com>
CC: Raphael S. Carvalho <raphaelsc@scylladb.com>